Feat: 가운데 basket에 받은 덕담들 표시 적용
